Refactor tests and remove deprecated global setup
- Deleted `setupGlobals.js` and moved necessary global shims directly into test files.
- Removed `smoke.js` and consolidated its tests into more specific test files.
- Deleted old test files for `studentSubmission` and `taskDefinition`, replacing them with updated versions.
- Added new test files for `assignment`, `artifacts`, `phase1Models`, and `imageManager` with comprehensive test cases.
- Updated `vitest.config.js` to include new test directories.
- Enhanced test coverage for various models and request handlers, ensuring proper functionality and validation.
