+++ tests/models/artifacts.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { ArtifactFactory, TableTaskArtifact } from '../src/AdminSheet/Models/Artifacts/index.js';
+import { ArtifactFactory, TableTaskArtifact } from '../../src/AdminSheet/Models/Artifacts/index.js';
 
 describe('Artifacts', () => {
   it('normalises text content and hashes immediately for reference/template', () => {
@@ -20,7 +20,7 @@ describe('Artifacts', () => {
   });
 
   it('canonicalises spreadsheet formulas and hashes immediately', () => {
-    const ss = ArtifactFactory.spreadsheet({ type: 'spreadsheet', taskId: 't3', role: 'reference', content: [ ['=sum(a1:a2)','"a"'], ['=if("b",1,2)'] ] });
+    const ss = ArtifactFactory.spreadsheet({ type: 'spreadsheet', taskId: 't3', role: 'reference', content: [ ['=sum(a1:a2)', '"a"'], ['=if("b",1,2)'] ] });
     expect(ss.content[0][0]).toBe('=SUM(A1:A2)');
     expect(ss.content[0][1]).toBe('"a"');
     expect(ss.contentHash).toBeTruthy();
